Creating update_cd callback
